+++ cpp/open3d/t/geometry/kernel/PointCloudCUDA.cu
@@ -25,4 +25,4 @@
 // ----------------------------------------------------------------------------
 
 #include "open3d/core/kernel/CUDALauncher.cuh"
-#include "open3d/t/geometry/kernel/PointCloudShared.h"
+#include "open3d/t/geometry/kernel/PointCloudImpl.h"
